[IGNORE] move appendIfMissing to pkg config
